Delete original lynepad and replace it with lynepad2
