+++ keyboards/qvex/lynepad/config.h
@@ -1,11 +1,10 @@
 /*
-Copyright 2020 KemoNine
+Copyright 2023 QVEX Tech
 
 This program is free software: you can redistribute it and/or modify
 it under the terms of the GNU General Public License as published by
 the Free Software Foundation, either version 2 of the License, or
 (at your option) any later version.
-
 This program is distributed in the hope that it will be useful,
 but WITHOUT ANY WARRANTY; without even the implied warranty of
 MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
@@ -17,11 +16,20 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
 #pragma once
 
-/* Definitions for encoder tilt/press support */
-#define ENC_TILT_THRESHOLD 1
-#define PIN_TW_SW D2 // Center
-#define PIN_RJ_SW C6 // Center
-#define PIN_RJ_DIR_A D4 // Up
-#define PIN_RJ_DIR_B D7 // Left
-#define PIN_RJ_DIR_C B6 // Down
-#define PIN_RJ_DIR_D B4 // Right
+#define DYNAMIC_KEYMAP_LAYER_COUNT 5
+
+//Digital joystick (4-way switch + encoder)
+#define PIN_JU D4
+#define PIN_JD B6
+#define PIN_JL D7
+#define PIN_JR B4
+#define PIN_JC C6
+
+//Thumbwheel
+#define PIN_TC D2
+
+//Backlighting
+#define BACKLIGHT_DEFAULT_LEVEL 5
+
+//RGB arc
+#define RGBLIGHT_DEFAULT_VAL 64
