--- keyboards/qvex/lynepad/keymaps/default/readme.md
@@ -1,1 +1,0 @@
-# The default keymap for lynepad
