--- keyboards/qvex/lynepad2/config.h
@@ -1,35 +1,0 @@
-/*
-Copyright 2023 QVEX Tech
-
-This program is free software: you can redistribute it and/or modify
-it under the terms of the GNU General Public License as published by
-the Free Software Foundation, either version 2 of the License, or
-(at your option) any later version.
-This program is distributed in the hope that it will be useful,
-but WITHOUT ANY WARRANTY; without even the implied warranty of
-MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-GNU General Public License for more details.
-
-You should have received a copy of the GNU General Public License
-along with this program.  If not, see <http://www.gnu.org/licenses/>.
-*/
-
-#pragma once
-
-#define DYNAMIC_KEYMAP_LAYER_COUNT 5
-
-//Digital joystick (4-way switch + encoder)
-#define PIN_JU D4
-#define PIN_JD B6
-#define PIN_JL D7
-#define PIN_JR B4
-#define PIN_JC C6
-
-//Thumbwheel
-#define PIN_TC D2
-
-//Backlighting
-#define BACKLIGHT_DEFAULT_LEVEL 5
-
-//RGB arc
-#define RGBLIGHT_DEFAULT_VAL 64
