--- keyboards/qvex/lynepad2/info.json
@@ -1,73 +1,0 @@
-{
-    "manufacturer": "QVEX Tech",
-    "keyboard_name": "Lynepad",
-    "maintainer": "QVEX-Tech",
-    "bootloader": "caterina",
-    "diode_direction": "COL2ROW",
-    "matrix_pins": {
-        "cols": [ "F0", "F1", "F4", "F5"],
-        "rows": [ "C7", "F7" ,"F6", null, null]
-    },
-    "processor": "atmega32u4",
-    "url": "https://qvex.eu",
-    "usb": {
-        "device_version": "1.0.0",
-        "vid": "0x04D8",
-        "pid": "0xE678"
-    },
-    "encoder": {
-        "rotary": [
-            {"pin_a": "D6", "pin_b": "B5", "resolution": 2},
-            {"pin_a": "D1", "pin_b": "D0", "resolution": 2}
-        ]
-    },
-    "features": {
-        "backlight": true,
-        "bootmagic": true,
-        "encoder": true,
-        "extrakey": true,
-        "mousekey": true,
-        "rgblight": true
-    },
-    "rgblight": {
-        "led_count": 5,
-        "max_brightness": 192,
-        "sleep": true
-    },
-    "ws2812": {
-        "pin": "D3"
-    },
-    "backlight": {
-        "breathing": true,
-        "levels": 8,
-        "max_brightness": 192,
-        "pin": "B7"
-    },
-    "layouts": {
-        "LAYOUT": {
-            "layout": [
-                { "matrix": [0, 0], "x": 0, "y": 0 },
-                { "matrix": [0, 1], "x": 1, "y": 0 },
-                { "matrix": [0, 2], "x": 2, "y": 0 },
-                { "matrix": [0, 3], "x": 3, "y": 0 },
-
-                { "matrix": [1, 0], "x": 0, "y": 1 },
-                { "matrix": [1, 1], "x": 1, "y": 1 },
-                { "matrix": [1, 2], "x": 2, "y": 1 },
-                { "matrix": [1, 3], "x": 3, "y": 1 },
-
-                { "matrix": [2, 0], "x": 0, "y": 2 },
-                { "matrix": [2, 1], "x": 1, "y": 2 },
-                { "matrix": [2, 2], "x": 2, "y": 2 },
-
-                { "matrix": [3, 0], "x": 4.75, "y": 1.3 },
-                { "matrix": [3, 1], "x": 4.75, "y": 3.7 },
-                { "matrix": [3, 2], "x": 3.55, "y": 2.5 },
-                { "matrix": [3, 3], "x": 5.95, "y": 2.5 },
-                { "matrix": [4, 0], "x": 4.55, "y": 2.3, "w": 1.4, "h": 1.4 },
-
-                { "matrix": [4, 1], "x": 3.4, "y": 5.2, "w": 1.6, "h": 1.6 }
-            ]
-        }
-    }
-}
