--- keyboards/qvex/lynepad2/rules.mk
@@ -1,2 +1,0 @@
-CUSTOM_MATRIX = lite
-SRC += matrix.c
